MIR5/atnf - just a simple copy actually

+++ src/subs/scrio.c
@@ -8,11 +8,12 @@
 /*   rjs  26oct95  Better messages on errors.				*/
 /*   pjt  19jun02  MIR4 prototypes                                      */
 /*   jwr  05nov04  Change file offsets to type off_t			*/
+/*   rjs  03jan05  Include file rationalisation.			*/
 /************************************************************************/
 
 #include <stdio.h>
-#include "io.h"
 #include "miriad.h"
+#include "io.h"
 
 
 static int number=0;
@@ -66,7 +67,7 @@ void scrclose_c(int handle)
   }
 }
 /************************************************************************/
-void scrread_c(int handle,float *buffer,int offset,int length)
+void scrread_c(int handle,float *buffer,off_t offset,size_t length)
 /**scrread -- Read real data from a scratch file.			*/
 /*:scratch-i/o								*/
 /*+  FORTRAN call sequence:
@@ -96,7 +97,7 @@ void scrread_c(int handle,float *buffer,int offset,int length)
   }
 }
 /************************************************************************/
-void scrwrite_c(int handle,Const float *buffer,int offset,int length)
+void scrwrite_c(int handle,Const float *buffer,off_t offset,size_t length)
 /**scrwrite -- Write real data to the scratch file.			*/
 /*:scratch-i/o								*/
 /*+  FORTRAN call sequence:
